+++ CHANGELOG.md
@@ -270,3 +270,4 @@ All notable changes to this project will be documented in this file.
 - remove meta_info from file intrinsics
 - add allow_import polyfill in file intrinsics
 - add default value info in hoverdocs
+- use react instead of plain html